Moved match carousel into shared ui.
