+++ android/app/src/main/java/com/adammcneilly/pocketleague/ui/FeedContent.kt
@@ -17,11 +17,11 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.adammcneilly.pocketleague.R
 import com.adammcneilly.pocketleague.android.designsystem.components.EmptyStateCard
-import com.adammcneilly.pocketleague.android.designsystem.matches.MatchesCarousel
 import com.adammcneilly.pocketleague.android.designsystem.matches.RecentMatchesEmptyState
 import com.adammcneilly.pocketleague.core.displaymodels.EventSummaryDisplayModel
 import com.adammcneilly.pocketleague.core.displaymodels.MatchDetailDisplayModel
 import com.adammcneilly.pocketleague.feature.feed.FeedViewState
+import com.adammcneilly.pocketleague.shared.ui.match.MatchCarousel
 import com.adammcneilly.pocketleague.ui.composables.eventsummary.EventSummaryListItem
 
 /**
@@ -113,7 +113,7 @@ private fun RecentMatchesSection(
     onMatchClicked: (String) -> Unit,
 ) {
     if (recentMatches.isNotEmpty()) {
-        MatchesCarousel(
+        MatchCarousel(
             matches = recentMatches,
             onMatchClicked = onMatchClicked,
         )
