+++ shared/ui/src/commonMain/kotlin/com/adammcneilly/pocketleague/shared/ui/match/MatchCard.kt
@@ -5,7 +5,6 @@ import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.text.InlineTextContent
 import androidx.compose.foundation.text.appendInlineContent
@@ -51,7 +50,6 @@ fun MatchCard(
             ),
     ) {
         Column(
-//            verticalArrangement = Arrangement.spacedBy(4.dp),
             modifier = Modifier
                 .padding(PocketLeagueTheme.sizes.cardPadding),
         ) {
